Stack Problems V1.5: Fibonacci logic fully working - all done on activity 'D' - still needs to work on aesthetical issues.

+++ src/dynamicstack.hpp
@@ -12,7 +12,7 @@ typedef struct Block Block;
 typedef struct DPilha DPilha;
 
 struct DItem{
-	int value;
+	long long unsigned int value;
     int index;
     string val;
 };
